Moved new Highline form out of bottom sheet
